+++ src/raid/raid1/bitmap.hpp
@@ -36,13 +36,13 @@ public:
     bool is_dirty(uint64_t addr, uint32_t len);
 
     // Tuple of form [page*, page_offset, size_consumed (max len)]
-    std::tuple< word_t*, uint32_t, uint32_t > dirty_page(uint64_t addr, uint32_t len);
+    std::tuple< word_t*, uint32_t, uint32_t > dirty_page(uint64_t addr, uint64_t len);
     std::tuple< word_t*, uint32_t, uint32_t > clean_page(uint64_t addr, uint32_t len);
     std::pair< uint64_t, uint32_t > next_dirty();
 
     // Each bit in the BITMAP represents a single "Chunk" of size chunk_size
     static std::tuple< uint32_t, uint32_t, uint32_t, uint32_t, uint64_t >
-    calc_bitmap_region(uint64_t addr, uint32_t len, uint32_t chunk_size);
+    calc_bitmap_region(uint64_t addr, uint64_t len, uint32_t chunk_size);
 
     void init_to(UblkDisk& device);
     void load_from(UblkDisk& device);
